pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "locus-photo"
version = "0.1.0"
description = "A photo viewer with GPS metadata editing and map integration."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0-or-later" }
authors = [{ name = "Kyrylo Protsenko" }]
keywords = ["photo", "viewer", "exif", "gps", "metadata", "raw"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Viewers",
]
dependencies = [
    "numpy",
    "rawpy",
    "pillow",
    "pyyaml",
    "PySide6",
    "geocoder",
    "pyperclip",
    "pyexiftool",
]

[project.optional-dependencies]
build = ["pyinstaller"]
dev = ["pytest", "pytest-cov"]

[project.urls]
Homepage = "https://github.com/kp9991-git/locus-photo"
Repository = "https://github.com/kp9991-git/locus-photo"
Issues = "https://github.com/kp9991-git/locus-photo/issues"

[tool.setuptools.packages.find]
include = ["viewer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
